+++ tools/themes/tool.js
@@ -8,7 +8,7 @@
         name: 'themes',
         category: 'appearance',
         type: 'multiple',
-        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'slimmerArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'pieceDrag','noPractice', 'gameMoveList', 'fatGauge', 'fatMove', 'gridBoard'/*, 'noGrab'*/],
+        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'slimmerArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'toggleStudyChat', 'pieceDrag','noPractice', 'gameMoveList', 'fatGauge', 'fatMove', 'gridBoard'],
         defaultValue: '',
         advanced: true
       }
@@ -27,8 +27,8 @@
         'themes.lessIcons': 'Less Icons',
         'themes.nonStickyHeader': 'No sticky header',
         'themes.noStudyChat': 'No study chat',
+        'themes.toggleStudyChat': 'Toggle study chat',
         'themes.pieceDrag': 'Nicer piece drag',
-        'themes.noGrab': 'No grab cursor',
         'themes.noPractice': 'No Practice button',
         'themes.gameMoveList': 'Flexible game move list',
         'themes.fatGauge': 'Thick analysis gauge',
@@ -48,8 +48,8 @@
         'themes.lessIcons': 'Mai pu\u0163ine pictograme',
         'themes.nonStickyHeader': 'Antet fix \u00een pagin\u0103',
         'themes.noStudyChat': 'F\u0103r\u0103 chat \u00een studii',
+        'themes.toggleStudyChat': 'Comut\u0103 chat \u00een studii',
         'themes.pieceDrag': 'Apucat piese mai frumos',
-        'themes.noGrab': 'F\u0103r\u0103 cursor care apuc\u0103',
         'themes.noPractice': 'F\u0103r\u0103 buton Antrenament cu calculatorul',
         'themes.gameMoveList': 'List\u0103 mut\u0103ri flexibil\u0103 \u00een joc',
         'themes.fatGauge': 'Bar\u0103 analiz\u0103 groas\u0103',
@@ -81,6 +81,7 @@
         const backgroundImage = styles.getPropertyValue('background-image');
         lt.global.document.documentElement.style.setProperty('--board-background', backgroundImage||'unset');
       }
+      $('body').toggleClass('lichessTools-hasBoardBackground', !!board.length);
     }
 
     applyThemes = ()=>{